+++ src/main/java/bootcampJava/EvaluacionFinal/repository/StockRepository.java
@@ -11,7 +11,7 @@ import bootcampJava.EvaluacionFinal.entity.StockId;
 public interface StockRepository extends JpaRepository<Stock, StockId> {
 	
 	@Query("SELECT s FROM Stock s WHERE s.idProducto = :idProducto ORDER BY s.stockDisponible DESC")
-	List<Stock> findTop3ByIdProductoOrderByStockDisponibleDesc(int idProducto, org.springframework.data.domain.Pageable pageable);
+	List<Stock> findTop3ByIdProductoOrderByStockDisponibleDesc(Long idProducto, org.springframework.data.domain.Pageable pageable);
     
-    List<Stock> findByIdSucursal(int idSucursal);
+    List<Stock> findByIdSucursal(Long idSucursal);
 }
